Reorganize
(also adds some placeholders for unimplemented color + hexadecimal libraries)
