+++ scripts/headers/Common/Convert/Hexadecimal.h
