+++ scripts/headers/Common/Output/FormattedDebug.h
